+++ R/netclu_greedy.R
@@ -29,7 +29,8 @@
 #' (`return_node_type = "both"` by default).
 #' 
 #' @param algorithm_in_output a `boolean` indicating if the original output
-#' of `communities` should be returned in the output (see Value).
+#' of `cluster_fast_greedy` should be returned in the output (`TRUE` by 
+#' default, see Value).
 #'
 #' @export
 #' @details
@@ -63,7 +64,7 @@
 #' \item{**clusters**: `data.frame` containing the clustering results}}
 #'
 #' In the `algorithm` slot, if `algorithm_in_output = TRUE`, users can
-#' find an "communities" object, output of
+#' find the output of
 #' [cluster_fast_greedy][igraph::cluster_fast_greedy].
 #'
 #' @author