+++ R/netclu_labelprop.R
@@ -28,7 +28,8 @@
 #' (`return_node_type = "both"` by default).
 #' 
 #' @param algorithm_in_output a `boolean` indicating if the original output
-#' of `communities` should be returned in the output (see Value).
+#' of `cluster_label_prop]` should be returned in the output (`TRUE` by 
+#' default, see Value).
 #'
 #' @details
 #' This function is based on propagating labels
@@ -61,7 +62,7 @@
 #' \item{**clusters**: `data.frame` containing the clustering results}}
 #'
 #' In the `algorithm` slot, if `algorithm_in_output = TRUE`, users can
-#' find an "communities" object, output of
+#' find a "communities" object, output of
 #' [cluster_label_prop][igraph::cluster_label_prop].
 #'
 #' @author