+++ R/similarity.R
@@ -7,14 +7,14 @@
 #' @param comat a co-occurrence `matrix` with sites as rows and species as
 #' columns.
 #' 
-#' @param metric a vector of string(s) indicating which metrics to chose
+#' @param metric a `character` vector indicating which metrics to chose
 #' (see Details). Available options are *abc*, *ABC*, *Jaccard*,
 #' *Jaccardturn*, *Sorensen*, *Simpson*,  *Bray*,
 #' *Brayturn* or *Euclidean*.\cr
 #' If `"all"` is specified, then all metrics will be
 #' calculated. Can be set to `NULL` if `formula` is used.
 #' 
-#' @param formula a vector of string(s) with your own formula based on the
+#' @param formula a `character` vector with your own formula(s) based on the
 #' `a`, `b`, `c`, `A`, `B`, and `C` quantities
 #' (see Details). `formula` is set to `NULL` by default.
 #' 
@@ -54,7 +54,7 @@
 #' `formula = c("1 - pmin(b,c) / (a + pmin(b,c))", "1 - (B + C) / (2*A + B + C)")`
 #' will compute the Simpson and Bray-Curtis similarity metrics, respectively. 
 #' **Note that pmin is used in the Simpson formula because a, b, c, A, B and C 
-#' are vectors.**
+#' are `numeric` vectors.**
 #'
 #' Euclidean computes the Euclidean similarity between each pair of site
 #' following this equation: